update name and test ownerOf on mint

+++ src/lib/BitPackedMap.sol
@@ -10,7 +10,11 @@ contract BitPackedMap {
     /// @notice Given a value for a square, return the corresponding fill rgb hex string.
     /// @param  square 4 bit uint denoting the color of the square.
     /// @return fill rgb hex string corresponding to square.
-    function _getFillFromSquare(uint256 square) internal pure returns (string memory) {
+    function _getFillFromSquare(uint256 square)
+        internal
+        pure
+        returns (string memory)
+    {
         string[16] memory colors = [
             "#ffffff",
             "#aaaaaa",
@@ -37,7 +41,11 @@ contract BitPackedMap {
     /// @param  bitmap 32 byte value containing all bitmap data.
     /// @param  index index of square being retrieved.
     /// @return square 4 bit uint denoting the color of the square.
-    function _getSquareFromMap(bytes32 bitmap, uint256 index) internal pure returns (uint256 square) {
+    function _getSquareFromMap(bytes32 bitmap, uint256 index)
+        internal
+        pure
+        returns (uint256 square)
+    {
         assembly {
             let shift := sub(252, mul(index, 4))
             square := shr(shift, and(shl(shift, shr(0xFC, not(0))), bitmap))
@@ -48,22 +56,99 @@ contract BitPackedMap {
     /// @param  bitmap 32 byte value containing all bitmap data.
     /// @return svg string to display bitmap.
     function _renderSvg(bytes32 bitmap) internal pure returns (string memory) {
-        string memory svgString = '<?xml version="1.0" encoding="UTF-8" standalone="no"?><svg xmlns="http://www.w3.org/2000/svg" version="1.1" viewBox="0 0 8 8">';
+        string memory svgString =
+            '<?xml version="1.0" encoding="UTF-8" standalone="no"?><svg xmlns="http://www.w3.org/2000/svg" version="1.1" viewBox="0 0 8 8">';
 
         string[8] memory pixelRow;
 
         uint256 y = 0;
-        for (uint i; i < 64; i += 8) {
-            pixelRow[0] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i)), '" x="0" y="', _toString(y), '" width="1" height="1" />'));
-            pixelRow[1] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i + 1)), '" x="1" y="', _toString(y), '" width="1" height="1" />'));
-            pixelRow[2] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i + 2)), '" x="2" y="', _toString(y), '" width="1" height="1" />'));
-            pixelRow[3] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i + 3)), '" x="3" y="', _toString(y), '" width="1" height="1" />'));
-            pixelRow[4] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i + 4)), '" x="4" y="', _toString(y), '" width="1" height="1" />'));
-            pixelRow[5] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i + 5)), '" x="5" y="', _toString(y), '" width="1" height="1" />'));
-            pixelRow[6] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i + 6)), '" x="6" y="', _toString(y), '" width="1" height="1" />'));
-            pixelRow[7] = string(abi.encodePacked('<rect fill="', _getFillFromSquare(_getSquareFromMap(bitmap, i + 7)), '" x="7" y="', _toString(y), '" width="1" height="1" />'));
-            
-            svgString = string(abi.encodePacked(svgString, pixelRow[0], pixelRow[1], pixelRow[2], pixelRow[3], pixelRow[4], pixelRow[5], pixelRow[6], pixelRow[7]));
+        for (uint256 i; i < 64; i += 8) {
+            pixelRow[0] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i)),
+                    '" x="0" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+            pixelRow[1] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i + 1)),
+                    '" x="1" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+            pixelRow[2] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i + 2)),
+                    '" x="2" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+            pixelRow[3] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i + 3)),
+                    '" x="3" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+            pixelRow[4] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i + 4)),
+                    '" x="4" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+            pixelRow[5] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i + 5)),
+                    '" x="5" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+            pixelRow[6] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i + 6)),
+                    '" x="6" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+            pixelRow[7] = string(
+                abi.encodePacked(
+                    '<rect fill="',
+                    _getFillFromSquare(_getSquareFromMap(bitmap, i + 7)),
+                    '" x="7" y="',
+                    _toString(y),
+                    '" width="1" height="1" />'
+                )
+            );
+
+            svgString = string(
+                abi.encodePacked(
+                    svgString,
+                    pixelRow[0],
+                    pixelRow[1],
+                    pixelRow[2],
+                    pixelRow[3],
+                    pixelRow[4],
+                    pixelRow[5],
+                    pixelRow[6],
+                    pixelRow[7]
+                )
+            );
 
             unchecked {
                 ++y;
@@ -91,28 +176,28 @@ contract BitPackedMap {
     // SOURCE: https://etherscan.io/address/0x8d04a8c79ceb0889bdd12acdf3fa9d207ed3ff63#code
     function _uintToHexDigit(uint8 d) private pure returns (bytes1) {
         if (0 <= d && d <= 9) {
-            return bytes1(uint8(bytes1('0')) + d);
+            return bytes1(uint8(bytes1("0")) + d);
         } else if (10 <= uint8(d) && uint8(d) <= 15) {
-            return bytes1(uint8(bytes1('a')) + d - 10);
+            return bytes1(uint8(bytes1("a")) + d - 10);
         }
         revert();
     }
 
     // SOURCE: https://etherscan.io/address/0x8d04a8c79ceb0889bdd12acdf3fa9d207ed3ff63#code
-    function _uintToHexString(uint a) private pure returns (string memory) {
-        uint count = 0;
-        uint b = a;
+    function _uintToHexString(uint256 a) private pure returns (string memory) {
+        uint256 count = 0;
+        uint256 b = a;
         while (b != 0) {
             count++;
             b /= 16;
         }
         bytes memory res = new bytes(count);
-        for (uint i=0; i<count; ++i) {
+        for (uint256 i = 0; i < count; ++i) {
             b = a % 16;
             res[count - i - 1] = _uintToHexDigit(uint8(b));
             a /= 16;
         }
-        
+
         string memory str = string(res);
         if (bytes(str).length == 0) {
             return "00";
